rename newPromptsParser to promptFileParser, move out of service folder
