+++ src/vs/workbench/contrib/chat/browser/promptSyntax/promptFileRewriter.ts
@@ -10,7 +10,7 @@ import { EditOperation } from '../../../../../editor/common/core/editOperation.j
 import { Range } from '../../../../../editor/common/core/range.js';
 import { ITextModel } from '../../../../../editor/common/model.js';
 import { ILanguageModelToolsService, IToolAndToolSetEnablementMap } from '../../common/languageModelToolsService.js';
-import { PromptHeaderAttributes } from '../../common/promptSyntax/service/newPromptsParser.js';
+import { PromptHeaderAttributes } from '../../common/promptSyntax/promptFileParser.js';
 import { IPromptsService } from '../../common/promptSyntax/service/promptsService.js';
 
 export class PromptFileRewriter {
