+++ src/vs/workbench/contrib/chat/browser/promptSyntax/promptToolsCodeLensProvider.ts
@@ -20,7 +20,7 @@ import { registerEditorFeature } from '../../../../../editor/common/editorFeatur
 import { PromptFileRewriter } from './promptFileRewriter.js';
 import { Range } from '../../../../../editor/common/core/range.js';
 import { IEditorModel } from '../../../../../editor/common/editorCommon.js';
-import { PromptHeaderAttributes } from '../../common/promptSyntax/service/newPromptsParser.js';
+import { PromptHeaderAttributes } from '../../common/promptSyntax/promptFileParser.js';
 
 class PromptToolsCodeLensProvider extends Disposable implements CodeLensProvider {
