+++ src/vs/workbench/contrib/chat/test/common/mockPromptsService.ts
@@ -10,7 +10,7 @@ import { URI } from '../../../../../base/common/uri.js';
 import { ITextModel } from '../../../../../editor/common/model.js';
 import { IExtensionDescription } from '../../../../../platform/extensions/common/extensions.js';
 import { PromptsType } from '../../common/promptSyntax/promptTypes.js';
-import { ParsedPromptFile } from '../../common/promptSyntax/service/newPromptsParser.js';
+import { ParsedPromptFile } from '../../common/promptSyntax/promptFileParser.js';
 import { ICustomAgent, IPromptPath, IPromptsService, PromptsStorage } from '../../common/promptSyntax/service/promptsService.js';
 
 export class MockPromptsService implements IPromptsService {
